refactor: 코드 정리 및 외부 기술 infra 이전
