+++ build.gradle
@@ -31,6 +31,10 @@ dependencies {
 	implementation 'org.springframework.boot:spring-boot-starter-aop'
 	implementation 'org.springframework.boot:spring-boot-starter'
 
+	//open feign
+	implementation group: 'org.springframework.cloud', name: 'spring-cloud-starter-openfeign', version: '4.1.3'
+	implementation 'org.springframework.cloud:spring-cloud-commons:4.1.4'
+
 	//ELK 연동을 위한 외부 기술
 	implementation 'org.springframework.boot:spring-boot-starter-data-elasticsearch'
 	implementation 'net.logstash.logback:logstash-logback-encoder:7.0' // Logstash용 로그백 엔코더
