--- src/main/java/com/petmatz/api/example/TestDTO.java
@@ -1,11 +1,0 @@
-package com.petmatz.api.example;
-
-import lombok.Data;
-
-@Data
-public class TestDTO {
-
-    private String profileImage;
-    private String standard;
-
-}
