--- src/main/java/com/petmatz/api/example/TestRequest.java
@@ -1,11 +1,0 @@
-package com.petmatz.api.example;
-
-import com.petmatz.domain.example.TestDTO;
-import lombok.Data;
-
-public record TestRequest(
-        String base64Image
-) {
-
-
-}
