+++ src/main/java/com/petmatz/api/global/dto/S3ImgDataResponse.java
@@ -3,7 +3,6 @@ package com.petmatz.api.global.dto;
 import com.petmatz.domain.global.S3ImgDataInfo;
 import lombok.Builder;
 
-@Builder
 public record S3ImgDataResponse(
 
         Long UUID,
@@ -12,10 +11,7 @@ public record S3ImgDataResponse(
 
 ) {
     public static S3ImgDataResponse of(S3ImgDataInfo petSaveInfo) {
-        return S3ImgDataResponse.builder()
-                .UUID(petSaveInfo.UUID())
-                .resultImgURL(petSaveInfo.resultImgURL())
-                .build();
+        return new S3ImgDataResponse(petSaveInfo.UUID(), petSaveInfo.resultImgURL());
     }
 
 }
