+++ src/main/java/com/petmatz/api/pet/PetController.java
@@ -4,14 +4,16 @@ import com.petmatz.api.global.dto.Response;
 
 import com.petmatz.api.global.dto.S3ImgDataResponse;
 import com.petmatz.api.pet.dto.PetApiRequest;
-import com.petmatz.api.pet.dto.PetInfoDto;
+import com.petmatz.api.pet.dto.PetInfoResponse;
 import com.petmatz.api.pet.dto.PetRequest;
 import com.petmatz.api.pet.dto.PetUpdateRequest;
 import com.petmatz.common.security.utils.JwtExtractProvider;
 import com.petmatz.domain.global.S3ImgDataInfo;
 import com.petmatz.domain.pet.PetService;
-import com.petmatz.domain.pet.vo.PetInf;
-import com.petmatz.domain.pet.vo.PetUpdateInfo;
+import com.petmatz.domain.pet.dto.OpenApiPetInfo;
+import com.petmatz.domain.pet.dto.PetInf;
+import com.petmatz.domain.pet.dto.PetUpdateInfo;
+import com.petmatz.domain.user.component.UserReader;
 import com.petmatz.domain.user.entity.User;
 import com.petmatz.domain.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
@@ -28,8 +30,9 @@ import java.net.MalformedURLException;
 public class PetController {
 
     private final PetService petService;
-    //TODO 바꿔야 됨
+    //TODO 바꿔야 됨 -> userReader -> userService로
 //    private final UserService userService;
+    private final UserReader userReader;
 
     private final UserRepository userRepository;
     private final JwtExtractProvider jwtExtractProvider;
@@ -37,10 +40,10 @@ public class PetController {
     // 동물등록번호 조회
     @PostMapping("/fetch")
     @Operation(summary = "동물등록번호 조회", description = "외부 API를 통해 동물등록번호 정보를 조회합니다.")
-    public Response<PetInfoDto> fetchPetInfo(@RequestBody PetApiRequest request) {
+    public Response<PetInfoResponse> fetchPetInfo(@RequestBody PetApiRequest request) {
         PetInf serviceDto = PetApiRequest.toServiceDto(request);
-        PetInf fetchedServiceDto = petService.fetchPetInfo(serviceDto.dogRegNo(), serviceDto.ownerNm());
-        PetInfoDto responseDto = PetInfoDto.of(fetchedServiceDto);
+        OpenApiPetInfo openApiPetInfo = petService.fetchPetInfo(serviceDto.dogRegNo(), serviceDto.ownerNm());
+        PetInfoResponse responseDto = PetInfoResponse.of(openApiPetInfo);
         return Response.success(responseDto);
     }
 
@@ -48,7 +51,8 @@ public class PetController {
     @PostMapping("/register")
     @Operation(summary = "반려동물 등록", description = "사용자의 반려동물 정보를 등록합니다.")
     public Response<S3ImgDataResponse> registerPet(@RequestBody PetRequest request) throws MalformedURLException {
-        User user = getAuthenticatedUser();
+        Long userId = jwtExtractProvider.findIdFromJwt();
+        User user = userReader.getAuthenticatedUser(userId);
         S3ImgDataInfo petSaveInfo = petService.savePet(user, PetInf.of(request));
         return Response.success(S3ImgDataResponse.of(petSaveInfo));
     }
@@ -58,7 +62,8 @@ public class PetController {
     @Operation(summary = "반려동물 정보 수정", description = "기존 반려동물 정보를 수정합니다.")
     @Parameter(name = "id", description = "반려동물 ID", example = "1")
     public Response<S3ImgDataResponse> updatePet(@PathVariable Long id, @RequestBody PetUpdateRequest petUpdateRequest) throws MalformedURLException {
-        User user = getAuthenticatedUser();
+        Long userId = jwtExtractProvider.findIdFromJwt();
+        User user = userReader.getAuthenticatedUser(userId);
         S3ImgDataInfo petSaveInfo = petService.updatePet(id, user, PetUpdateInfo.of(petUpdateRequest));
         return Response.success(S3ImgDataResponse.of(petSaveInfo));
     }
@@ -68,16 +73,14 @@ public class PetController {
     @Operation(summary = "반려동물 삭제", description = "등록된 반려동물을 삭제합니다.")
     @Parameter(name = "id", description = "반려동물 ID", example = "1")
     public Response<Void> deletePet(@PathVariable Long id) {
-        User user = getAuthenticatedUser();
+        Long userId = jwtExtractProvider.findIdFromJwt();
+        User user = userReader.getAuthenticatedUser(userId);
         petService.deletePet(id, user);
         return Response.success("댕댕이 정보가 성공적으로 삭제되었습니다.");
     }
 
-    private User getAuthenticatedUser() {
-            Long userId = jwtExtractProvider.findIdFromJwt();
-            return userRepository.findById(userId)
-                    .orElseThrow(() -> new IllegalArgumentException("User not found :" + userId));
-    }
+
+
 }
