+++ src/main/java/com/petmatz/api/sosboard/SosBoardController.java
@@ -11,7 +11,6 @@ import com.petmatz.domain.sosboard.dto.SosBoardServiceDto;
 import com.petmatz.domain.user.entity.User;
 import com.petmatz.domain.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
