--- src/main/java/com/petmatz/domain/example/TestDTO.java
@@ -1,15 +1,0 @@
-package com.petmatz.domain.example;
-
-import lombok.Builder;
-import lombok.Data;
-
-
-@Builder
-public record TestDTO(
-
-        String name,
-        String age
-
-) {
-
-}
