--- src/main/java/com/petmatz/domain/example/TestOpenApiInterface.java
@@ -1,7 +1,0 @@
-package com.petmatz.domain.example;
-
-public interface TestOpenApiInterface {
-
-    void getTest();
-
-}
