+++ src/main/java/com/petmatz/domain/pet/entity/Pet.java
@@ -3,8 +3,7 @@ package com.petmatz.domain.pet.entity;
 import com.petmatz.domain.global.BaseEntity;
 import com.petmatz.domain.pet.Gender;
 import com.petmatz.domain.pet.Size;
-import com.petmatz.domain.pet.vo.PetInf;
-import com.petmatz.domain.pet.vo.PetUpdateInfo;
+import com.petmatz.domain.pet.dto.PetUpdateInfo;
 import com.petmatz.domain.petmission.entity.PetToPetMissionEntity;
 import com.petmatz.domain.user.entity.User;
 import jakarta.persistence.*;
@@ -74,23 +73,6 @@ public class Pet extends BaseEntity {
 
     public void addUser(User userData) {
         user = userData;
-    }
-
-    public static Pet of(PetInf petInf, String imgURL, User user) {
-        return Pet.builder()
-                .user(user)
-                .dogRegNo(petInf.dogRegNo())
-                .petName(petInf.petName())
-                .breed(petInf.breed())
-                .gender(Gender.fromString(petInf.gender()))
-                .neuterYn(petInf.neuterYn())
-                .size(Size.fromString(petInf.size()))
-                .age(petInf.age())
-                .temperament(petInf.temperament())
-                .preferredWalkingLocation(petInf.preferredWalkingLocation())
-                .profileImg(imgURL)
-                .comment(petInf.comment())
-                .build();
     }
 
     public void updatePetInfo(PetUpdateInfo petUpdateInfo, String imgUrl) {
