--- src/main/java/com/petmatz/infra/example/TestOpenApi.java
@@ -1,13 +1,0 @@
-package com.petmatz.infra.example;
-
-import com.petmatz.domain.example.TestOpenApiInterface;
-import org.springframework.stereotype.Component;
-
-
-@Component
-public class TestOpenApi implements TestOpenApiInterface {
-    @Override
-    public void getTest() {
-
-    }
-}
